HR Management System/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
marshmallow==3.20.1
PyMySQL==1.1.0